Update LICENSE to ensure we protect ourselves from others selling rotabot
The idea behind why we are doing rotabot is to ensure I can develop this idea and other can use it. But I'm not really liking the idea of others using my work and sell it as a service.

All the code for rotabot will be available so if you want to run it yourselves for your employes please go ahead! There will be tools available for you to migrate from the cloud rotabot to your own and viceversa.