docs: Update LICENSE to Elastic License 2.0 (ELv2) (#9)
The idea behind why we are doing rotabot is to ensure I can develop this idea and others can use it. But I'm not really liking the idea of others using my work and selling it as a service.

All the code for rotabot will be available so if you want to run it yourselves for your employer please go ahead! There will be tools available for you to migrate from the cloud rotabot to your own and vice-versa.

Choosing ELv2 since this is the best version for the above similar to what [artie](https://artie.so/) does