feat: Allow users to update a rota
This pull requests enhances the "add_rota" view to make it handle
creation or updates or the rota. I initially tried to do the same with
two different views but it did not make much sense as most of the code
was the same.

Since we can know which button the user clicked and understand whether
they want to create or update we show the same view with different calls
to actions (create vs update).

Part of this PR was developed in a GoCardless hackathon with my
colleagues, I just brought the ideas of the hackathon to reality with
this PR (and also the tests xD)

Co-authored-by: Hendrickson Langbehn <hendricksonrl@gmail.com>
Co-authored-by: Greg Kelly <gkelly@gocardless.com>