feat: Automatically update updated_at on change (#13)
We could do this ourselves on the code but I feel staying away of "DB
Magic" at all cost is a bit of an extreme. This allows us to safely keep
the updated_at up to date without needing to send it all the time (and
miss it sometimes). Since I don't expect Rotabot to be a DB intense
application where we update millions of resources at once I feel this is
fine.